Got setEnvVars() in utility lib working
